split seeds into multiple files
